+++ Domain/Repositories/BccrCurrencyRepositoryUtils.cs
@@ -44,7 +44,7 @@ namespace Domain.Repositories
             return new BccrSingleVentanillaModel
             {
                 LastUpdate = lastUpdate,
-                Value = value
+                Value = (decimal)value
             };
         }
     }
